+++ packages/dql/src/Functions/Property.php
@@ -17,7 +17,7 @@ class Property extends \EDT\Querying\Functions\Property implements ClauseFunctio
         return [];
     }
 
-    public function asDql(array $valueReferences, array $propertyAliases): string
+    public function asDql(array $valueReferences, array $propertyAliases, string $mainEntityAlias): string
     {
         Assert::count($valueReferences, 0);
         Assert::count($propertyAliases, 1);